Move split_chunks() into slicetools
